LayoutTests/fast/viewport/viewport-41.html

<head>
    <title>Viewport META with "width=1000, height=device-height, initial-scale=0.5".</title>
    <meta name="viewport" content="width=1000, height=device-height, initial-scale=0.5">
    <script>
        function test() {
            if (window.layoutTestController) {
                layoutTestController.dumpAsText();
                layoutTestController.dumpConfigurationForViewport(320, 352);
            }
        }
    </script>
</head>
<body onload="test();">
