LayoutTests/fast/viewport/viewport-41-expected.txt

CONSOLE MESSAGE: line 3: Viewport width or height set to physical device width, try using "device-width" constant instead for future compatibility.
viewport size 1000x704 scale 0.500000 with limits [0.500000, 5.000000]
